Replace libc imports with std::ffi where trivial

+++ src/codec/subtitle/mod.rs
@@ -12,9 +12,9 @@ use std::mem;
 
 use crate::ffi::AVSubtitleType::*;
 use crate::ffi::*;
-use libc::{c_uint, size_t};
 #[cfg(feature = "serialize")]
 use serde::{Deserialize, Serialize};
+use std::ffi::c_uint;
 
 #[derive(Eq, PartialEq, Clone, Copy, Debug)]
 #[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
@@ -107,11 +107,10 @@ impl Subtitle {
             self.0.num_rects += 1;
             self.0.rects = av_realloc(
                 self.0.rects as *mut _,
-                (mem::size_of::<*const AVSubtitleRect>() * self.0.num_rects as usize) as size_t,
+                mem::size_of::<*const AVSubtitleRect>() * self.0.num_rects as usize,
             ) as *mut _;
 
-            let rect =
-                av_mallocz(mem::size_of::<AVSubtitleRect>() as size_t) as *mut AVSubtitleRect;
+            let rect = av_mallocz(mem::size_of::<AVSubtitleRect>()) as *mut AVSubtitleRect;
             (*rect).type_ = kind.into();
 
             *self.0.rects.offset((self.0.num_rects - 1) as isize) = rect;
